Hotel deletion now deletes the reservations, rooms, hotelprofile and reviews

+++ SleepGoBE/SleepGo/SleepGo.App/Features/Users/Commands/RemoveUserCommand.cs
@@ -2,6 +2,8 @@
 using Microsoft.Extensions.Logging;
 using SleepGo.App.Exceptions;
 using SleepGo.App.Interfaces;
+using SleepGo.Domain.Enums;
+using System.Runtime.InteropServices;
 
 namespace SleepGo.App.Features.Users.Commands
 {
@@ -27,30 +29,83 @@ namespace SleepGo.App.Features.Users.Commands
                 throw new UserNotFoundException($"The user with ID {request.userId} doesn't exist and it could not be removed!");
             }
 
-            var reviews = await _unitOfWork.ReviewRepository.GetAllReviewsByUserIdAsync(userToRemove.Id);
-            if(reviews != null)
-            {
-                foreach(var review in reviews)
-                {
-                    await _unitOfWork.ReviewRepository.RemoveAsync(review);
 
-                    //Add a recalculating hotel rating once adding a field.
+            if(userToRemove.Role == Role.User)
+            {
+                var reviews = await _unitOfWork.ReviewRepository.GetAllReviewsByUserIdAsync(userToRemove.Id);
+                if (reviews != null)
+                {
+                    foreach (var review in reviews)
+                    {
+                        await _unitOfWork.ReviewRepository.RemoveAsync(review);
+
+                        //Add a recalculating hotel rating once adding a field.
+                    }
+                }
+
+                var reservations = await _unitOfWork.ReservationRepository.GetAllReservationsByUserIdAsync(userToRemove.Id);
+                if (reservations != null)
+                {
+                    foreach (var reservation in reservations)
+                    {
+                        await _unitOfWork.ReservationRepository.RemoveAsync(reservation);
+
+                        var room = await _unitOfWork.RoomRepository.GetByIdAsync(reservation.RoomId);
+                        if (room != null)
+                        {
+                            room.Reservations = room.Reservations.Where(r => r.Id != reservation.Id).ToList();
+                            await _unitOfWork.RoomRepository.UpdateAsync(room);
+                        }
+                    }
                 }
             }
-
-            var reservations = await _unitOfWork.ReservationRepository.GetAllReservationsByUserIdAsync(userToRemove.Id);
-            if(reservations != null)
+            
+            if(userToRemove.Role == Role.Hotel)
             {
-                foreach(var reservation in reservations)
+                //var hotelToRemove = await _unitOfWork.HotelRepository.GetByIdAsync(userToRemove.Id);
+                var hotelToRemove = await _unitOfWork.HotelRepository.GetHotelProfileByUserId(userToRemove.Id);
+
+                if (hotelToRemove != null) 
                 {
-                    var room = await _unitOfWork.RoomRepository.GetByIdAsync(reservation.RoomId);
-                    if(room != null)
+                    var rooms = await _unitOfWork.RoomRepository.GetRoomsByHotelIdAsync(hotelToRemove.Id);
+                    if(rooms != null)
                     {
-                        room.Reservations.Remove(reservation);
-                        await _unitOfWork.RoomRepository.UpdateAsync(room);
+                        foreach(var room in rooms)
+
+                        {
+                            var roomReservations = await _unitOfWork.ReservationRepository.GetReservationsByRoomIdAsync(room.Id);
+
+                            if(roomReservations != null)
+                            {
+                                foreach(var res in roomReservations)
+                                {
+                                    await _unitOfWork.ReservationRepository.RemoveAsync(res);
+                                }
+                            }
+
+                            await _unitOfWork.RoomRepository.RemoveAsync(room);
+                        }
                     }
 
-                    await _unitOfWork.ReservationRepository.RemoveAsync(reservation);
+                    var amenities = await _unitOfWork.AmenityRepository.GetHotelAmenitiesByHotelIdAsync(hotelToRemove.Id);
+                    if (amenities != null)
+                    {
+                        foreach (var amenity in amenities)
+                        {
+                            await _unitOfWork.AmenityRepository.RemoveAsync(amenity);
+                        }
+                    }
+
+                    var hotelReviews = await _unitOfWork.ReviewRepository.GetAllReviewsByHotelIdAsync(hotelToRemove.Id);
+                    if(hotelReviews != null)
+                    {
+                        foreach(var review in hotelReviews)
+                        {
+                            await _unitOfWork.ReviewRepository.RemoveAsync(review);
+                        }
+                    }
+
+                    await _unitOfWork.HotelRepository.RemoveAsync(hotelToRemove);
                 }
             }
